Refactor patina_performance per crate organization guidelines
Updates the content of `patina_performance` to follow the most recent
organization format as defined in RFC 9.

https://github.com/OpenDevicePartnership/patina/blob/main/docs/src/rfc/text/0009-standardize-component-crate.md#standard-crate-layout
