+++ components/patina_performance/src/component.rs
@@ -1,0 +1,13 @@
+//! Patina Performance Components
+//!
+//! ## License
+//!
+//! Copyright (C) Microsoft Corporation. All rights reserved.
+//!
+//! SPDX-License-Identifier: BSD-2-Clause-Patent
+//!
+
+pub mod performance;
+
+// Re-export the Performance component for easier access.
+pub use performance::Performance;
