+++ docs/src/components/patina_performance.md
@@ -29,14 +29,14 @@ The Patina performance component uses a feature mask in its configuration to con
 
 Core::default()
  // ...
- .with_config(patina_performance::EnabledMeasurement(&[
-        patina_sdk::performance::measurement::Measurement::DriverBindingStart, // Adds driver binding start measurements.
-        patina_sdk::performance::measurement::Measurement::DriverBindingStop, // Adds driver binding stop measurements.
-        patina_sdk::performance::measurement::Measurement::DriverBindingSupport, // Adds driver binding support measurements.
-        patina_sdk::performance::measurement::Measurement::LoadImage, // Adds load image measurements.
-        patina_sdk::performance::measurement::Measurement::StartImage, // Adds start image measurements.
+ .with_config(patina_performance::config::EnabledMeasurement(&[
+        patina_sdk::performance::Measurement::DriverBindingStart,   // Adds driver binding start measurements.
+        patina_sdk::performance::Measurement::DriverBindingStop,    // Adds driver binding stop measurements.
+        patina_sdk::performance::Measurement::DriverBindingSupport, // Adds driver binding support measurements.
+        patina_sdk::performance::Measurement::LoadImage,            // Adds load image measurements.
+        patina_sdk::performance::Measurement::StartImage,           // Adds start image measurements.
     ]))
- .with_component(patina_performance::Performance)
+ .with_component(patina_performance::component::Performance))
  .start()
  .unwrap();
